pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codesearch"
version = "0.1.0"
description = "Semantic code search: BM25 vs dense vs hybrid — with rigorous eval"
requires-python = ">=3.11"
dependencies = [
    "rank-bm25>=0.2.2",
    "sentence-transformers>=3.0.0",
    "qdrant-client>=1.9.0",
    "gradio>=4.36.0",
    "datasets>=2.20.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/codesearch"]

[tool.ruff]
line-length = 88
src = ["src"]
